fixed cloud deployment and migrations for deleted seqr.VariantAnnotation model

+++ deploy/kubernetes/scripts/deploy_matchbox.sh
@@ -2,32 +2,32 @@
 
 
 SCRIPT_DIR="$( cd "$(dirname "$0")" ; pwd -P )"
-source ${SCRIPT_DIR}/check_env.sh
-
+source ${SCRIPT_DIR}/init_env.sh
 set -x
 
-# delete any previous deployments
-kubectl delete -f configs/matchbox/matchbox-deployment.yaml
-kubectl delete -f configs/matchbox/matchbox-service.yaml
-
-FORCE_ARG=
-if [ "$FORCE" = true ]; then
-    FORCE_ARG=--no-cache
+if [ "$DELETE_BEFORE_DEPLOY" ]; then
+    # delete any previous deployments
+    kubectl delete -f configs/matchbox/matchbox-deployment.yaml
+    kubectl delete -f configs/matchbox/matchbox-service.yaml
 fi
 
-docker build $FORCE_ARG -t ${DOCKER_IMAGE_PREFIX}/matchbox docker/matchbox/
+BUILD_ARG=
+if [ "$BUILD" ]; then
+    BUILD_ARG=--no-cache
+fi
+
+docker build $BUILD_ARG -t ${DOCKER_IMAGE_PREFIX}/matchbox docker/matchbox/
 if [ "$DEPLOY_TO" = 'gcloud' ]; then
     gcloud docker -- push ${DOCKER_IMAGE_PREFIX}/matchbox
 fi
 
-kubectl create -f configs/matchbox/matchbox-deployment.yaml --record
-kubectl create -f configs/matchbox/matchbox-service.yaml --record
+MATCHBOX_POD_NAME=$( kubectl get pods -o=name | grep 'matchbox-' | cut -f 2 -d / | tail -n 1 )
+if [ "$MATCHBOX_POD_NAME" ] ; then
+    kubectl set image -f configs/matchbox/matchbox-deployment.${DEPLOY_TO}.yaml --record
+    #kubectl edit -f configs/matchbox/matchbox-service.yaml --record
+else
+    kubectl create -f configs/matchbox/matchbox-deployment.yaml --record
+    kubectl create -f configs/matchbox/matchbox-service.yaml --record
+fi
 
-# wait for pod to start
-set +x
-while [ ! "$( kubectl get pods | grep 'matchbox-' | grep Running)" ] || [ "$( kubectl get pods | grep 'matchbox-' | grep Terminating)" ]; do
-    echo $(date) - Waiting for matchbox pod to enter "Running" state. Current state is: "$( kubectl get pods | grep 'matchbox-' )"
-    sleep 5
-done
-echo $(date) - Success. Current state is: "$( kubectl get pods | grep 'matchbox-' )"
-set -x
+wait_until_pod_is_running matchbox
